Datasource: Refactor
